requirements.txt
coolname==1.1.0
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-core==2.0.0
google-cloud-storage==1.42.3
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0
